Sharing and My files screens bootstraped and many more...

+++ src/main/java/servlet/SharedFilesServlet.java
@@ -1,7 +1,7 @@
 package servlet;
 
-import database.user.ShareService;
-import database.user.UserService;
+import database.dao.ShareService;
+import database.entity.ShareLog;
 import servlet.common.ServletUtils;
 
 import javax.servlet.ServletException;
@@ -11,14 +11,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 import java.util.Map;
 
 @WebServlet("/received")
 public class SharedFilesServlet extends HttpServlet {
 
-    private Map<Long, String> sharedFilesList;
+    private List<ShareLog> sharedFilesList;
     private String SHARE_DIRECTORY = "D:/apache-tomcat-9.0.12/uploads/share";
-    // private String SHARE_DIRECTORY = "/usr/local/apache-tomcat-9.0.12/uploads/share";
+//    private String SHARE_DIRECTORY = "/usr/local/apache-tomcat-9.0.12/uploads/share";
 
 
     @Override
@@ -31,7 +32,13 @@ public class SharedFilesServlet extends HttpServlet {
             if (ServletUtils.isAuthenticated(request)) {
                 ShareService shareService = new ShareService(SHARE_DIRECTORY);
                 String currentUser = (String) request.getSession().getAttribute("username");
-                sharedFilesList = shareService.getSharedFilesListByUsername(currentUser);
+
+                if (request.getParameter("searchWord") != null) {
+                    sharedFilesList = shareService.getSharedFilesBySearch(currentUser, request.getParameter("searchWord"));
+                } else {
+                    sharedFilesList = shareService.getSharedFilesListByUsername(currentUser);
+                }
+
                 request.setAttribute("sharedFiles", sharedFilesList);
                 request.getRequestDispatcher("/WEB-INF/jsps/views/received.jsp").forward(request, response);
                 return;
@@ -64,7 +71,9 @@ public class SharedFilesServlet extends HttpServlet {
             ex.printStackTrace();
         }
         ServletUtils.sendResponseFile(response, fileToSend);
-        fileToSend.delete();
+        if(!"dont".equals(decryptParameter)) {
+            fileToSend.delete();
+        }
     }
